ai-backend/requirements.txt
fastapi
uvicorn
ultralytics
pillow
python-multipart 
gitpython
matplotlib
numpy
opencv-python
Pillow
psutil
PyYAML
requests
scipy
thop
torch
torchvision
tqdm
pandas
seaborn
setuptools
